fix(schema): GWT JsInterop 숫자 타입 불일치로 인한 ClassCastException 해결
- 공유 도메인 모델(Type, TypeLayout)의 rev 필드 타입을 long에서 double로 변경하여 자바스크립트 number 타입과의 호환성 확보
- GWT 환경에서 JS number를 Java long으로 캐스팅할 때 발생하는 ClassCastException 원천 차단
- 백엔드(type-command, type-query)의 R2DBC 엔티티 변환 로직에서 double과 Long 간의 캐스팅 처리 추가
- 통합 테스트(R2dbcTypeRepositoryIntegrationTest)의 검증 및 날짜 범위 로직을 변경된 규격에 맞춰 업데이트

+++ schema/src/main/java/dev/sayaya/handbook/domain/Type.java
@@ -25,7 +25,7 @@ public final class Type {
     @JsonProperty("width") @JsProperty(name = "width") private double width;
     @JsonProperty("height") @JsProperty(name = "height") private double height;
     @JsonProperty("attributes") @JsProperty(name = "attributes") private Attribute[] attributes;
-    @JsonProperty("rev") @JsProperty(name = "rev") private long rev;
+    @JsonProperty("rev") @JsProperty(name = "rev") private double rev;
 
     @JsOverlay @JsIgnore
     public static Type create(String id, String version, double effectDateTime, double expireDateTime) {
@@ -34,7 +34,7 @@ public final class Type {
         type.version(version);
         type.effectDateTime(effectDateTime);
         type.expireDateTime(expireDateTime);
-        type.rev(-1);
+        type.rev(-1.0);
         return type;
     }
 
@@ -45,7 +45,7 @@ public final class Type {
         type.version(version);
         type.width(width);
         type.height(height);
-        type.rev(-1);
+        type.rev(-1.0);
         return type;
     }
