+++ type-command/src/main/kotlin/dev/sayaya/handbook/interfaces/database/R2dbcTypeEntity.kt
@@ -39,7 +39,7 @@ data class R2dbcTypeEntity(
             expireDateTime.toEpochMilli().toDouble()
         ).description(description).primitive(primitive).attributes(attributes.toTypedArray())
         parent?.let { type.parent(it) }
-        rev?.let { type.rev(it) }
+        rev?.let { type.rev(it.toDouble()) }
         return type
     }
 
@@ -53,7 +53,7 @@ data class R2dbcTypeEntity(
             description = type.description(),
             primitive = type.primitive(),
             parent = type.parent(),
-            rev = if (type.rev() == -1L) null else type.rev(),
+            rev = if (type.rev() == -1.0) null else type.rev().toLong(),
         )
     }
 }
